+++ src/modules/pedido/pedido.service.ts
@@ -1,13 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePedidoDto } from './dto/create-pedido.dto';
 import { UpdatePedidoDto } from './dto/update-pedido.dto';
-import { Pago } from 'src/entities/pago.entity';
 import { Cadete } from 'src/entities/cadete.entity';
 import { Cliente } from 'src/entities/cliente.entity';
 import { Producto } from 'src/entities/producto.entity';
 import { Pedido } from 'src/entities/pedido.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+
 
 @Injectable()
 export class PedidoService {
@@ -16,37 +16,41 @@ export class PedidoService {
     @InjectRepository(Producto) private productoRepository: Repository<Producto>,
     @InjectRepository(Cliente) private clienteRepository: Repository<Cliente>,
     @InjectRepository(Cadete) private cadeteRepository: Repository<Cadete>,
-    @InjectRepository(Pago) private pagoRepository: Repository<Pago>,
   ) {}
 
   async create(dto: CreatePedidoDto) {
     const producto = await this.productoRepository.findOneBy({ id: dto.producto });
-    const cliente = await this.clienteRepository.findOneBy({ id: dto.cliente });
-    const cadete = await this.cadeteRepository.findOneBy({ id: dto.cadete });
 
-    if (!producto || !cliente) throw new NotFoundException('Producto o Cliente no encontrado');
-    if (!cadete) throw new NotFoundException('Cadete no encontrado');
+    if (!producto) throw new NotFoundException('Producto no encontrado');
 
     const nuevoPedido = new Pedido();
+
+    if (dto.cliente) {
+      const cliente = await this.clienteRepository.findOneBy({ id: dto.cliente });
+      if (!cliente) throw new NotFoundException('Cliente no encontrado');
+      nuevoPedido.cliente = cliente;
+    }
+    if (dto.cadete) {
+      const cadete = await this.cadeteRepository.findOneBy({ id: dto.cadete });
+      if (!cadete) throw new NotFoundException('Cadete no encontrado');
+      nuevoPedido.cadete = cadete;
+    }
     nuevoPedido.producto = producto;
-    nuevoPedido.cliente = cliente;
-    nuevoPedido.cadete = cadete;
     nuevoPedido.cantidad = dto.cantidad;
-    nuevoPedido.total = dto.total;
-    nuevoPedido.fecha = dto.fecha;
+    nuevoPedido.total = producto.precio;
     nuevoPedido.estado = 'PENDIENTE';
-    nuevoPedido.observacion = dto.observacion;
+    if (dto.observacion !== undefined) nuevoPedido.observacion = dto.observacion;
 
     return this.pedidoRepository.save(nuevoPedido);
   }
 
   findAll() {
-    return this.pedidoRepository.find({ relations: ['producto', 'cliente', 'cadete', 'pago'] });
+    return this.pedidoRepository.find({ relations: ['producto', 'cliente', 'cadete'] });
   }
 
-  findOne(id: number) {
-    const pedido = this.pedidoRepository.findOne({ where: { id }, 
-                        relations: ['producto', 'cliente', 'cadete', 'pago'] });
+  async findOne(id: number) {
+    const pedido = await this.pedidoRepository.findOne({ where: { id }, 
+                        relations: ['producto', 'cliente', 'cadete'] });
     if (!pedido) {
       throw new NotFoundException('Pedido no encontrado');
     }
@@ -57,20 +61,28 @@ export class PedidoService {
     const pedido = await this.findOne(id);
     if (!pedido) throw new NotFoundException('Pedido no encontrado');
 
-    const producto = await this.productoRepository.findOneBy({ id: dto.producto });
-    const cliente = await this.clienteRepository.findOneBy({ id: dto.cliente });
-    const cadete = await this.cadeteRepository.findOneBy({ id: dto.cadete });
+    if (dto.producto) {
+      const producto = await this.productoRepository.findOneBy({ id: dto.producto });
+      if (!producto) throw new NotFoundException('Producto no encontrado');
+      pedido.producto = producto;
+    }
 
-    Object.assign(pedido, {
-      producto,
-      cliente,
-      cadete,
-      cantidad: dto.cantidad,
-      total: dto.total,
-      fecha: dto.fecha,
-      estado: dto.estado,
-      observacion: dto.observacion
-    });
+    if (dto.cliente) {
+      const cliente = await this.clienteRepository.findOneBy({ id: dto.cliente });
+      if (!cliente) throw new NotFoundException('Cliente no encontrado');
+      pedido.cliente = cliente;
+    }
+
+    if (dto.cadete) {
+      const cadete = await this.cadeteRepository.findOneBy({ id: dto.cadete });
+      if (!cadete) throw new NotFoundException('Cadete no encontrado');
+      pedido.cadete = cadete;
+    }
+
+    if (dto.cantidad !== undefined) pedido.cantidad = dto.cantidad;
+    if (dto.total !== undefined) pedido.total = pedido.producto.precio * pedido.cantidad;
+    if (dto.estado) pedido.estado = dto.estado;
+    if (dto.observacion !== undefined) pedido.observacion = dto.observacion;
 
     return this.pedidoRepository.save(pedido);
   }
@@ -78,17 +90,7 @@ export class PedidoService {
   async remove(id: number) {
     const pedido = await this.findOne(id);
     if (!pedido) throw new NotFoundException('Pedido no encontrado.');
+    if (pedido.estado == "PAGADO") throw new BadRequestException('No se puede eliminar un pedido pagado');
     return this.pedidoRepository.remove(pedido);
-  }
-
-  // Asociar un pago existente a un pedido
-  async asignarPago(pedidoId: number, pagoId: number) {
-    const pedido = await this.pedidoRepository.findOneBy({ id: pedidoId });
-    const pago = await this.pagoRepository.findOneBy({ id: pagoId });
-
-    if (!pedido || !pago) throw new NotFoundException('Pedido o pago no encontrado');
-    pedido.pago = pago;
-
-    return this.pedidoRepository.save(pedido);
   }
 }
